backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-multipart==0.0.9
edge-tts==6.1.12
openai-whisper==20231117
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
peft>=0.11.0
httpx==0.27.0
python-dotenv==1.0.1
